cpu: x64: ukernel: split API and abstractions into different files
