+++ src/cpu/ukernel/transform.cpp
@@ -19,7 +19,7 @@
 #include "cpu/ukernel/c_types_map.hpp"
 
 #if DNNL_X64
-#include "cpu/x64/ukernel/brgemm_api.hpp"
+#include "cpu/x64/ukernel/transform.hpp"
 #endif
 
 #ifdef DNNL_EXPERIMENTAL_UKERNEL
